backend/requirements.txt
flask
flask-cors
pymongo
PyJWT
python-dotenv
werkzeug
dnspython
